moved aug utils to one file
